pyproject.toml
[project]
name = "langgraph-multiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromedriver-autoinstaller>=0.6.4",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "ipython>=9.4.0",
    "langchain-community>=0.3.28",
    "langchain-google-community>=2.0.7",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
    "langsmith>=0.4.19",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20250506",
    "pillow>=11.3.0",
    "playwright>=1.54.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "selenium>=4.35.0",
    "streamlit>=1.49.0",
    "web-extract-data>=0.1.0",
]
